Add "public_actions" to Interface DSL; sessions are not required for this.
Declaring public actions in any interface within a single service application will make that application less efficient, as ApiTools has to run through as far as interface selection before it can abandon processing should a session not be defined. Normally, it can halt processing *much* earlier if session data is missing.

In addition, the X-Service-Response-Time header is added and quotes a benchmark on the dispatch to a service.

Test coverage coming soon.