Fix up various tests, so all pass; but no new coverage for new features, yet.
Note that the "dup self and invoke _call" dance is now removed, since the middleware no longer uses instance variables. A future refactor might be to implement most of the functionality in class methods, with as much offloaded to e.g. the Interaction class as possible where it makes sense (many internal methods are now only passed the 'interaction' value as a parameter - which looks a bit like they should be "interaction.<method>", rather than "middleware.<method>(interaction)".