Fix some RDoc formatting and content issues.
The use of "+" for monospaced text is limited and didn't work in one case - wrapped in <tt>...</tt> instead. While using my own documents as a reference I noticed an issue where example text was being partially parsed as part of a following code sample due to indentation; fixed; also noticed no information on simple value types in Hashes, so added a "limitations" section to explain this.