Pointless change to work around apparent GitHub bug
Previous commit was pushed with a CI skip instruction for Travis intentionally. This worked, but GitHub then locked out the PR in a "waiting for results of tests" state which it would never resolve. Online suggestions included closing and re-opening the PR, or using the "Test service" button in the Travis integration for the project. None of these worked, so I'm forced to use the ugly hack of a commit just to trigger a build on something that's had no code changes.