DSL extension for "simple" types in Hash keys.
Work in progress; single test failure in a "keys" case for Array types still
to fix. Intend also to extend this to Array definitions too, so an Array of
simple types can be described.

Unavoidable syntax limitation requires simple field options for "keys" to be
prefixed with "field_" to avoid collision of e.g. the "length" option, for
the "keys" key names, with the "length" option of an associated simple String
type. In this case, "field_length" would refer to the String value.

[ci skip]