Fix for Hoodoo::Client.
When errors were present in the response that were *not* "invalid session", Client would nonetheless reacquire a session and retry (typically getting the same error back the second time too). This is obviously inefficient and risky (it might cause side effects twice despite the error, depending on the nature of the API being called).