Resolves issue #212, for the most part.
It's "for the most part" because the deprecation of the clunky "#bypass" mechanism in the old mock storage system isn't restored here, so the behaviour does change. At least the automatic use of a mock backend when in test mode with no Memcached host is now (re)implemented, which should resolve the test failures in services that were assuming automatic mock backend use; the work-around of manually selecting a mock backend in e.g. the RSpec before-suite configuration is no longer necessary.