Second cut of the Travis workaround.
The previous commit resulted in tests passing, but was a weak test. This tries to dig a little deeper and stub Timeout#timeout itself, checking specifically for parameters matching those given at the Hoodoo API layer and for the Net::OpenTimeout exception requested by Net::HTTP itself. We then simulate an immediate timeout by raising the requested exception. This test makes similar levels of assumption to, and is as strong as, the original TCPSocket#open stub which failed on Travis.