Stop using a deprecated API.
The dating system used #sanitize, which in AR 4 was present but not
considered part of the public API. It was entirely removed in AR 5.
For the use cases at hand, using #connection/#quoted_date is a more
semantically meaningful API anyway, albeit with slightly different
output behaviour; switch to that, so that tests under AR 5 pass.