Make test more robust.
An estimated-count test needed a relatively undisturbed table in
order to get the pre-and-post expected estimated counts it wanted
from PostgreSQL. Some seeds would disturb this by prior tests in
the same file inserting rows.

To fix this, the tests are split up into two overall contexts and
the outermost 'describe' requests ordered execution. Within that,
the first context does the ANALYZE-based fragile tests in order.
After that, the rest of the tests are executed randomly.