[skip ci] Realise test coverage was insufficient.
The case of "historic records exist but no contemporary does" was not covered
and upon realising that, I realised that the manual dating mechanism would fail
under the current implementation (and indeed it does). As the lightbulbs came
on, it became obvious that the 'undated scope' approach is untenable (if you
use manual dating you require a dated scope at all times since contemporary and
historic records exist within the same table). There was also the wrinkle of
'acquire_in' being slowed down by additional checks unnecessarily if dating was
not enabled. The intent is to fix this in a subsequent commit but having the
dating models officially monkey patch 'acquire_in' and after calling 'super',
check using module-appropriate contemporary record lookup if need be.