Handle pre ruby 2.4 Fixnums and Floats
Previous versions of Ruby would raise an exception to calling #dup on some types of Objects such as Fixnum, Bignum, and Float.

Unfortunately using #respond_to?(:dup) would not work as it would return true due to the implementation of their exception handling. Therefore a rescue seems like the simplest and most readable solution, for what is hopefully a temporary patch.