Clarified misleading comment after PR feedback.
The use of "tmpnam" not "tmpname" was intentional as I just assumed
that Ruby's implementation would use the C/C++ "tmpnam" function for
the implementation. Surprisingly, the diff for the Ruby core change
set shows it never did and the Rails solution is basically a copy
of the Ruby core stuff.